pyproject.toml
[project]
name = "rectoform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "easyofd==0.5.1",
    "httpx>=0.28.1",
    "openai>=2.28.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "rq>=2.6.0",
    "uvicorn>=0.35.0",
    "pytest>=9.0.2",
    "fastapi>=0.135.2",
    "pymupdf>=1.27.2",
]
